UI Card detail: Externalize card response to another component
