+++ ui/main/src/app/modules/cards/components/card-reponse/card-response.component.html
@@ -1,0 +1,44 @@
+<!-- Copyright (c) 2022, RTE (http://www.rte-france.com)              -->
+<!-- See AUTHORS.txt                                                       -->
+<!-- This Source Code Form is subject to the terms of the Mozilla Public   -->
+<!-- License, v. 2.0. If a copy of the MPL was not distributed with this   -->
+<!-- file, You can obtain one at http://mozilla.org/MPL/2.0/.              -->
+<!-- SPDX-License-Identifier: MPL-2.0                                      -->
+
+
+<button id="opfab-card-details-btn-response" class="opfab-btn"
+    *ngIf='showButton && isUserEnabledToRespond && !isResponseLocked'
+    [disabled]='lttdExpiredIsTrue || !isUserEnabledToRespond || sendingResponseInProgress'
+    (click)='processClickOnSendResponse()'>{{btnValidateLabel|translate}}</button>
+
+<button id="opfab-card-details-btn-response" class="opfab-btn" [disabled]='lttdExpiredIsTrue || !isUserEnabledToRespond'
+    *ngIf='showButton && isUserEnabledToRespond && isResponseLocked'
+    (click)='unlockAnswer()'>{{btnUnlockLabel|translate}}</button>
+
+
+<ng-template #chooseEntitiesForResponsePopup let-modal>
+    <div class="modal-header">
+        <div translate> response.entitiesChoice </div>
+        <div class="opfab-close-modal-icon" aria-label="Close" (click)="modal.dismiss('Cross click')">
+            <span aria-hidden="true">&times;</span>
+        </div>
+    </div>
+    <div class="modal-body text-center">
+        <br />
+        <p translate>response.chooseEntitiesForResponse</p><br />
+        <form [formGroup]="selectEntitiesForm" class="opfab-form-lighter">
+            <div class="row">
+                <of-multi-select id="opfab-card-details-entities-choice-selector" multiSelectId="entities"
+                    [parentForm]="selectEntitiesForm" [config]="multiSelectConfig"
+                    [options]="userEntityOptionsDropdownList"
+                    [selectedOptions]="selectEntitiesForm.get('entities').value">
+                </of-multi-select>
+            </div>
+            <br />
+        </form>
+        <button id="opfab-card-details-entity-choice-btn-cancel" type="button" class="opfab-btn-cancel"
+            (click)="cancelEntitiesChoice()" translate>button.cancel</button>
+        <button id="opfab-card-details-entity-choice-btn-confirm" type="button" class="opfab-btn"
+            (click)="submitEntitiesChoice()" translate>button.ok</button>
+    </div>
+</ng-template>
